conformance/buffimgs/cl_context_createBuffer.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLContext createBuffer conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCLContext :: createBuffer test case.");
    debug("(CLenum memFlags, CLuint sizeInBytes, optional ArrayBuffer hostPtr)");
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var hostPtr = new ArrayBuffer(SIZE);

    // Testing for MEM_READ_ONLY Buffer.
    shouldBeType("webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE);", "WebCLBuffer");
    shouldBeType("webCLContext.createBuffer(webcl.MEM_READ_ONLY, DATA_SIZE, hostPtr);", "WebCLBuffer");

    // Testing for MEM_WRITE_ONLY Buffer.
    shouldBeType("webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE);", "WebCLBuffer");
    shouldBeType("webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, DATA_SIZE, hostPtr);", "WebCLBuffer");

    // Testing for MEM_READ_WRITE Buffer.
    shouldBeType("webCLContext.createBuffer(webcl.MEM_READ_WRITE, DATA_SIZE);", "WebCLBuffer");
    shouldBeType("webCLContext.createBuffer(webcl.MEM_READ_WRITE, DATA_SIZE, hostPtr);", "WebCLBuffer");

    // Testing for Exceptions.
    shouldThrowWebCLException("webCLContext.createBuffer(invalid_object, DATA_SIZE);");
    shouldThrowWebCLException("webCLContext.createBuffer(webcl.MEM_READ_WRITE, invalid_object);");
    shouldThrowWebCLException("webCLContext.createBuffer(webcl.MEM_READ_WRITE, DATA_SIZE, invalid_object);");
    shouldThrowWebCLException("webCLContext.createBuffer(webcl.MEM_READ_WRITE, 0, hostPtr);");
    shouldThrowWebCLException("webCLContext.createBuffer(webcl.MEM_READ_WRITE, 100, new ArrayBuffer(10));");

    var webCLPlatforms = wtu.getPlatforms();
    var webCLDevices = wtu.getDevices(webCLPlatforms);
    var deviceMaxMem = webCLDevices[0].getInfo(webcl.DEVICE_MAX_MEM_ALLOC_SIZE);
    var deviceMaxMemSize = deviceMaxMem + 1;
    shouldThrowWebCLException("webCLContext.createBuffer(webcl.MEM_READ_ONLY, deviceMaxMemSize, new ArrayBuffer(1024));");

} catch(e) {
    testFailed("WebCLContext :: createBuffer() throws an exception : " + e.message);
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
